activateViewpoint() sets orthogonal and perspective
pI.activateViewpoint() now fully supports the setting of the orthogonal viewpoints and the perspective viewpoints.

+++ bcfplugin/rdwr/viewpoint.py
@@ -225,7 +225,7 @@ class OrthogonalCamera(Camera, XMLName):
             viewPoint: Point,
             direction: Direction,
             upVector: Direction,
-            viewWorldScale: int,
+            viewWorldScale: float,
             containingElement = None,
             state: State.States = State.States.ORIGINAL):
